add new higher-order classes
